In the purpose of democratizing Twitter's sorting algorithm, I had an impartial AI (ChatGPT 3) rewrite it in a completely uinbiased way
Here's the new twitter algorithm, as provided by a fully unbiased and impartial system. Hopefully Elon can fire some more of those pesky humans soon.

---

To sort tweets by relevance to a user, we first need to define what "relevance" means in this context. One possible definition is the degree to which a tweet is related to topics that the user is interested in. We can use various techniques to determine the relevance of a tweet, such as keyword matching, sentiment analysis, and user interaction history.

Assuming that we have a function `getRelevance(tweet: Tweet, user: User): Double` that returns the relevance score of a tweet to a user, we can sort a list of tweets by relevance using the `sortBy` method in Scala. Here's an example implementation:

```scala
case class Tweet(text: String, author: String, timestamp: Long)
case class User(interests: List[String])

def getRelevance(tweet: Tweet, user: User): Double = {
  // TODO: Implement relevance scoring algorithm
  0.0
}

def sortTweetsByRelevance(tweets: List[Tweet], user: User): List[Tweet] = {
  tweets.sortBy(tweet => -getRelevance(tweet, user))
}

// Example usage:
val tweets = List(
  Tweet("I love dogs", "Alice", 1648777200),
  Tweet("Just ate a delicious pizza", "Bob", 1648777300),
  Tweet("Breaking news: Earthquake in Japan", "CNN", 1648777400),
  Tweet("Check out my new blog post about gardening", "Alice", 1648777500)
)
val user = User(List("dogs", "gardening"))
val relevantTweets = sortTweetsByRelevance(tweets, user)
println(relevantTweets)
```

In this example, the `sortTweetsByRelevance` function takes a list of tweets and a user object, and returns a new list of tweets sorted by relevance to the user. The `getRelevance` function takes a tweet and a user, and returns a relevance score as a double. The tweets are sorted in descending order of relevance by passing a negated relevance score to the `sortBy` method. Finally, we print the sorted list of tweets using the `println` method.

Note that the `getRelevance` function is left unimplemented in this example, as the exact algorithm will depend on the specific requirements of the application.