Update ExternalDataSources.scala
I noticed that the "Interested in" recommendations had a lookback window of 14 days. I regularly use the "Not interested in this tweet" functionality, but notice topics I'm not interested in still seem to appear at times. I've modified the lookback window of the "Not interested in" topics to have the same 14-day lookback window for both consistency and perhaps improving on this functionality as users' dislikes are unlikely to change that quickly.