Correct improper promotion of low quality content
Turns out being able to pay to boost your content to the top of replies results in the people with the worst possible replies signing up. It's the only way they get seen because they're so low quality the algorithm naturally drops them out of view. So now that we have a simple way of identifying the lowest quality replies it's simple to change them to be forced to the bottom instead of the top. (I would have changed the scale parameters but I couldn't find them defined in this repo. But this should save looking up the parameters resulting in better algorithm performance.)